+++ package.json
@@ -1,12 +1,20 @@
 {
   "name": "suyotechdb",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "description": "JSON File Based Database Inspired by Mongoose ODM",
+  "version": "0.0.2",
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build" : "node build.mjs"
   },
-  "keywords": [],
-  "author": "suyotech",
-  "license": "ISC"
+  "repository": {
+    "url": "https://github.com/suyotech/suyotechdb.git"
+  },
+  "files": ["/dist","README.md"],
+  "keywords": ["embeddable","lightweight","json","file based","db","database","mongoose","odm","suyotechdb","suyotech"],
+  "author": "suyotech.com",
+  "license": "ISC",
+  "devDependencies": {
+    "esbuild": "^0.23.0"
+  }
 }
