+++ src/index.mjs
@@ -1,0 +1,6 @@
+
+import Schema from "./schema.mjs";
+import Model from "./model.mjs";
+
+
+export {Schema,Model}
